Updated for current API fields
The API supports a 'PUT' for modifying parts of an asset, but most of the supported fields were missing (I need to be able to set a location for example which was missing).  I've added all the current API fields and replaced the $values section with Get-ParameterValue.  Not fully tested everything but looking good so far.